+++ integration-test/src/test/java/org/apache/iotdb/db/it/IoTDBLoadTsFileIT.java
@@ -21,8 +21,6 @@ package org.apache.iotdb.db.it;
 
 import org.apache.iotdb.commons.auth.entity.PrivilegeType;
 import org.apache.iotdb.commons.schema.column.ColumnHeaderConstant;
-import org.apache.iotdb.db.storageengine.dataregion.compaction.tablemodel.CompactionTableModelTestFileWriter;
-import org.apache.iotdb.db.storageengine.dataregion.tsfile.TsFileResource;
 import org.apache.iotdb.it.env.EnvFactory;
 import org.apache.iotdb.it.framework.IoTDBTestRunner;
 import org.apache.iotdb.it.utils.TsFileGenerator;
@@ -33,10 +31,8 @@ import org.apache.iotdb.itbase.env.BaseEnv;
 import org.apache.iotdb.jdbc.IoTDBSQLException;
 
 import org.apache.tsfile.enums.TSDataType;
-import org.apache.tsfile.file.metadata.enums.CompressionType;
 import org.apache.tsfile.file.metadata.enums.TSEncoding;
 import org.apache.tsfile.read.common.Path;
-import org.apache.tsfile.read.common.TimeRange;
 import org.apache.tsfile.utils.Pair;
 import org.apache.tsfile.write.record.Tablet;
 import org.apache.tsfile.write.schema.IMeasurementSchema;
@@ -498,75 +494,6 @@ public class IoTDBLoadTsFileIT {
 
     executeNonQuery(
         String.format("load \"%s\" sgLevel=2", tmpDir.getAbsolutePath()), "test", "test123");
-  }
-
-  @Test
-  public void testTableAuth() throws Exception {
-    createUser("test", "test123");
-
-    final TsFileResource resource4 = new TsFileResource(new File(tmpDir, "test1-0-0-0.tsfile"));
-    try (final CompactionTableModelTestFileWriter writer =
-        new CompactionTableModelTestFileWriter(resource4)) {
-      writer.registerTableSchema("t2", Arrays.asList("id1", "id2", "id3"));
-      writer.startChunkGroup("t2", Arrays.asList("id_field1", "id_field2", "id_field3"));
-      writer.generateSimpleNonAlignedSeriesToCurrentDevice(
-          "s1",
-          new TimeRange[][][] {new TimeRange[][] {new TimeRange[] {new TimeRange(20, 22)}}},
-          TSEncoding.PLAIN,
-          CompressionType.LZ4);
-      writer.generateSimpleNonAlignedSeriesToCurrentDevice(
-          "s2",
-          new TimeRange[][][] {new TimeRange[][] {new TimeRange[] {new TimeRange(20, 22)}}},
-          TSEncoding.PLAIN,
-          CompressionType.LZ4);
-      writer.endChunkGroup();
-      writer.endFile();
-    }
-
-    try (final Connection userCon =
-            EnvFactory.getEnv().getConnection("test", "test123", BaseEnv.TABLE_SQL_DIALECT);
-        final Statement userStmt = userCon.createStatement()) {
-      Assert.assertThrows(
-          SQLException.class,
-          () -> {
-            userStmt.execute(
-                String.format(
-                    "load '%s' with ('database-level'='2', 'database-name'='test')",
-                    tmpDir.getAbsolutePath()));
-          });
-    }
-
-    try (Connection adminCon = EnvFactory.getEnv().getConnection(BaseEnv.TABLE_SQL_DIALECT);
-        Statement adminStmt = adminCon.createStatement()) {
-      adminStmt.execute("grant create on database test to user test");
-    }
-
-    try (final Connection userCon =
-            EnvFactory.getEnv().getConnection("test", "test123", BaseEnv.TABLE_SQL_DIALECT);
-        final Statement userStmt = userCon.createStatement()) {
-      Assert.assertThrows(
-          SQLException.class,
-          () -> {
-            userStmt.execute(
-                String.format(
-                    "load '%s' with ('database-level'='2', 'database-name'='test')",
-                    tmpDir.getAbsolutePath()));
-          });
-    }
-
-    try (final Connection adminCon = EnvFactory.getEnv().getConnection(BaseEnv.TABLE_SQL_DIALECT);
-        final Statement adminStmt = adminCon.createStatement()) {
-      adminStmt.execute("grant insert on any to user test");
-    }
-
-    try (final Connection userCon =
-            EnvFactory.getEnv().getConnection("test", "test123", BaseEnv.TABLE_SQL_DIALECT);
-        final Statement userStmt = userCon.createStatement()) {
-      userStmt.execute(
-          String.format(
-              "load '%s' with ('database-level'='2', 'database-name'='test')",
-              tmpDir.getAbsolutePath()));
-    }
   }
 
   @Test
