+++ iotdb-core/datanode/src/test/java/org/apache/iotdb/db/storageengine/dataregion/compaction/tablemodel/CompactionTableModelTestFileWriter.java
@@ -61,6 +61,20 @@ public class CompactionTableModelTestFileWriter extends CompactionTestFileWriter
     schema.registerTableSchema(tableSchema);
   }
 
+  public void registerTableSchemaWithTagField(String tableName, List<String> columnNames) {
+    List<IMeasurementSchema> measurementSchemas = new ArrayList<>();
+    List<ColumnCategory> columnTypes = new ArrayList<>();
+    int cnt = 0;
+    for (String columnName : columnNames) {
+      measurementSchemas.add(
+          new MeasurementSchema(
+              columnName, TSDataType.STRING, TSEncoding.PLAIN, CompressionType.UNCOMPRESSED));
+      columnTypes.add(cnt % 2 == 0 ? ColumnCategory.FIELD : ColumnCategory.TAG);
+      cnt++;
+    }
+    schema.registerTableSchema(new TableSchema(tableName, measurementSchemas, columnTypes));
+  }
+
   public IDeviceID startChunkGroup(String tableName, List<String> idColumnFields)
       throws IOException {
     int idx = 0;
